add LPFileTranslator
