+++ test/translators/resources/test_quadratic_program.lp
